# docs/swagger.yaml
basePath: /
definitions:
...
  product.Product:
    properties:
      calorie:
        type: string
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      time:
        type: string
      updated_at:
        type: string
      url_img:
        type: string
    type: object
  product.RequestProduct:
    properties:
      calorie:
        type: string
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      time:
        type: string
      url_img:
        type: string
    type: object
  product.RequestUpdateProduct:
    properties:
      calorie:
        type: string
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      time:
        type: string
      url_img:
        type: string
    type: object
...
  product.ResponseAnny:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  product.ResponseProduct:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/product.Product'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  rol.RequestRol:
    properties:
      name:
        type: string
    type: object
  rol.RequestUpdateRol:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  rol.ResponseAnny:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  rol.ResponseRol:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/rol.Rol'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  rol.Rol:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: 172.147.77.149:5050
info:
  contact:
    email: luis.lucero@bjungle.net
    name: API Support
  description: Api para la gestión de la comida y recetas
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: API FOOD
  version: "1.4"
paths:
...
  /api/v1/user/product:
    post:
      consumes:
      - application/json
      description: Método que permite crear un producto de comida
      parameters:
      - description: Datos de el producto
        in: body
        name: RequestProduct
        required: true
        schema:
          $ref: '#/definitions/product.RequestProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ResponseProduct'
      summary: Método que permite crear un producto de comida
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Método que permite actualizar un producto de comida
      parameters:
      - description: Datos del producto
        in: body
        name: RequestUpdateProduct
        required: true
        schema:
          $ref: '#/definitions/product.RequestUpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ResponseProduct'
      summary: Método que permite actualizar un producto de comida
      tags:
      - Product
  /api/v1/user/product/{id}:
    delete:
      consumes:
      - application/json
      description: Método que permite borrar un producto de comida
      parameters:
      - description: Id del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ResponseAnny'
      summary: Método que permite borrar un producto de comida
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Método que permite obtener un producto de comida por su id
      parameters:
      - description: Id del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ResponseProduct'
      summary: Método que permite obtener un producto de comida por su id
      tags:
      - Product
...
  /api/v1/user/rol:
    post:
      consumes:
      - application/json
      description: Método que permite crear un rol de comida
      parameters:
      - description: Datos del rol
        in: body
        name: RequestRol
        required: true
        schema:
          $ref: '#/definitions/rol.RequestRol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rol.ResponseRol'
      summary: Método que permite crear un rol de comida
      tags:
      - Rol
    put:
      consumes:
      - application/json
      description: Método que permite actualizar un rol de comida
      parameters:
      - description: Datos del rol
        in: body
        name: RequestUpdateRol
        required: true
        schema:
          $ref: '#/definitions/rol.RequestUpdateRol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rol.ResponseRol'
      summary: Método que permite actualizar un rol de comida
      tags:
      - Rol
  /api/v1/user/rol/{id}:
    delete:
      consumes:
      - application/json
      description: Método que permite borrar un rol de comida
      parameters:
      - description: Id del rol
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rol.ResponseAnny'
      summary: Método que permite borrar un rol de comida
      tags:
      - Rol
    get:
      consumes:
      - application/json
      description: Método que permite obtener un rol de comida por su id
      parameters:
      - description: Id del rol
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rol.ResponseRol'
      summary: Método que permite obtener un rol de comida por su id
      tags:
      - Rol
...
swagger: "2.0"
...